Add files via upload
We move from the features of the ERP characterized in Pereira (2015) towards the instantaneous amplitude and frequency of event-related changes correlated with a contrast in access and in phenomenology. We proceed as following. In the first section, empirical mode decomposition (EMD) with post processing (Xie et al. 2014) and Hilbert-Huang Transform (HHT). In the second section, calculated the variance inflation factor (VIF). In the third section, partial least squares regression (PLSR): the minimal root mean squared error of prediction (RMSEP). In the last section, partial least squares regression (PLSR): significance multivariate correlation (sMC) statistic.